+++ src/main/java/com/manchui/domain/repository/querydsl/ReviewQueryDslImpl.java
@@ -21,7 +21,7 @@ import java.util.Optional;
 import static com.manchui.domain.entity.QGathering.gathering;
 import static com.manchui.domain.entity.QImage.image;
 import static com.manchui.domain.entity.QReview.review;
-import static com.manchui.domain.entity.QUser.user;
+import static com.manchui.domain.user.entity.QUser.user;
 import static com.querydsl.jpa.JPAExpressions.select;
 
 @Slf4j
